requirements.txt
discord.py>=2.3.0
python-dotenv>=1.0.0
aiosqlite>=0.19.0
matplotlib>=3.8.0
pandas>=2.1.0
pillow>=10.0.0
aiofiles>=23.2.1
asyncio>=3.4.3
python-dateutil>=2.8.2